resources.conf: take out 10.42.42.1
This seems to be the default address used to communicate via SSH with the
sysmoBTS. Whichever process ends up getting this address sees all of the
SSH in its pcap (for the AoIP build it tends to be OsmoHLR).

We could filter properly, but actually also just take this address out of
the pool for allocation to server processes.

Change-Id: I07e74ba0b9a5b08a308aae7646c4b7c70fe4aa0e