refactor logging: drop 'with', simplify
Refactor logging to simplify use: drop the 'with Origin' style completely, and
instead use the python stack to determine which objects are created by which,
and which object to associate a log statement with.

The new way: we rely on the convention that each class instance has a local
'self' referencing the object instance. If we need to find an origin as a new
object's parent, or to associate a log message with, we traverse each stack
frame, fetching the first local 'self' object that is a log.Origin class
instance.

We just call log() and it finds an Origin object to log for.

When an exception happens, we first escalate the exception to where ever it is
handled, and only then the log.log_exn() function finds an Origin object in the
traceback's stack frames.

Hence the 'with log.Origin' happens implicitly, no more hassles.

Furthermore, any frame can place additional information in a local variable
called log_ctx. This automatically inserted in the ancestry associated with a
log statement.

Change-Id: I5f9b53150f2bb6fa9d63ce27f0806f0ca6a45e90