bts_osmotrx: Allow LD_LIBRARY_PATH when loading osmo-trx
When splitting osmo-trx to its own class the env injection part was
removed because it seemed not needed as it has no dependencies included
by default in the trial.

However, it may still be required by the osmo-gsm-tester main unit to
launch osmo-gsm-tester with LD_PRELOAD_PATH to point to specific
versions of libuhd, so let's add the env injection back in there.

Change-Id: I069ba69f4207665c8c575fed0fc8ff164c571e36