cosmetic: rename 'osmo-mgcpgw', it is inaccurate
The current binary is called osmo-bsc_mgcp, while the projected binary to be
used in the future will be called osmo-mgw. osmo-mgcpgw is a phantasy name that
never existed and never will.

Rename the suite.py "public API" to 'mgw', and use 'mgw' as name for
osmo-msc.cfg.

However, since currently the binary used is still osmo-bsc_mgcp, keep the parts
that are specific to this "MGW implementation" called osmo_bsc_mgcp /
OsmoBscMgcp. It is for itself 'osmo_bsc_mgcp', and for the MSC called 'mgw'.

Hence we can in the future add an osmo-mgw alongside the osmo-bsc_mgcp.

Change-Id: I9e2c485f0784c3b76e352fbb528ce17d1900f3d8